Enable Debugging with QEMU, display.h keyboard.h.
I used -s and -S flags to allow debugging with GDB, I just need to figure out the debug symbols in the binary with a hex editor, for the memory addresses.
I haven't been using a debugger because I couldn't figure out how to connect it to QEMU, but I really should've. I've added the MIT License, Just because.
Finally, the progress for the day. I've been working on the keyboard driver, for PS/2. It isn't working quite yet, but hopefully it will.
On the short-term to do-list:
1. Figure out debug symbols (Which GCC creates, I believe with the -g flag)
2. Finish the keyboard Driver. (I'm using a callback-based system, because I'm a JavaScript programmer. It might cause issues. We'll see)
3. Get the engine code running.