Commented code & tried to figure out debugger
Most of the day was spent trying to figure out GDB.
I commented a lot of my code so when I revisit it, it's easier to understand, and others will be able to read it too.
Most of my work was destroyed anyways. I'm running Ubuntu off a USB, meaning I have to reinstall everything.
It takes a few minutes, but I accidentally ran a command that rebooted automatically, and all my work was gone.
I was able to get back up to speed pretty fast though.